pyproject.toml
[tool.poetry]
name = "capitalisation_station"
version = "0.1.0"
description = ""
authors = ["eightballer"]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
cython = "<3.0.0"
open-aea-ledger-solana = "==1.52.0"
open-aea-ledger-cosmos = "==1.52.0"
open-aea-ledger-ethereum = "==1.52.0"
open-aea-cli-ipfs = "==1.52.0"
open-aea-test-autonomy = "==0.14.12"
open-autonomy = "==0.14.12"
autonomy-dev = {extras = ["all"], version = "0.2.43"}


[tool.poetry.dev-dependencies]


[tool.poetry.extras]
dev = ["pre-commit", "virtualenv", "pip", "toml", "bump2version"]
doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs",
]


[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "setuptools"]
